Copy BucketList.onMounted logic to post-onMounted bucketStore refreshes
This fixes a minor BucketList bug, where buckets only accessible via IDP perms would revert to dummy folders when folders are added/removed by the user.
This is because fetchBuckets() does not use the same logic to populate dummy folders with ones that may be accessible via IDP perms.

Also: in bucketStore, rename permResponse to userPermResponse for clarity (and to reflect difference between user perms and IDP perms).
